added routers to appjs
